Extract Position from the Strata-1 Beads
Was able to attempt extracting position of the beads by thresholding the image stack and then performing "bwlabeln" function to identify each bead. In order to ensure precise identification of the spherical beads, I had to place a 3D spherical erosion on the 3D image stack to remove briding pixels between differing beads.