Large and Medium Beads
Completed the extraction of the Position for the 10mm diameter spherical beads. There is a total of 180 10mm diameter beads and the erosion scan of 9 pixel radius is able to accurately exract 174/180. That is around a 96.67 % success rate. No false positives except for 3 pairs of conjoined 10mm beads. This means the remaining 6 beads are unidentified since each pair is seen as a single ID pair.